+++ lifecycle_test.go
@@ -1,0 +1,37 @@
+package fxt_test
+
+import (
+	"fmt"
+
+	"github.com/goph/fxt"
+	"go.uber.org/fx"
+)
+
+func ExampleLifecycle() {
+	var ctx struct {
+		Closer fxt.Closer
+	}
+
+	type A struct{}
+
+	fx.New(
+		fx.Provide(fxt.NewLifecycle),
+		fx.Provide(func(l fxt.Lifecycle) *A {
+			l.Append(fxt.Hook{
+				OnClose: func() error {
+					fmt.Print("closing")
+
+					return nil
+				},
+			})
+
+			return &A{}
+		}),
+		fx.Extract(&ctx),
+		fx.Invoke(func(a *A) {}),
+		fx.NopLogger,
+	)
+	defer ctx.Closer.Close()
+
+	// Output: closing
+}